fix: 花火が出る処理は別ファイルに切り出す
